Shrank trace on Net-(RN2-Pad1) for pin 13 of display
Pins 13, 14, 15, 16 on the LED segment shorted together on two boards AFTER soldering the display down. I believe that is because trace Net-(RN2-Pad1) from pin 13 on the alphanumeric display was too close to pins 14, 15, 16, resulting in a short after soldering and when the display is pressed down. I changed that trace to 6 mils (from 10 mils) and moved it away from the problem zone.